Changes:
Location.java: added a getter for address
res/*: I added a bunch of images that might or might not be needed, they
were in the open source project that I used
The BalloonItemizedOverlay and BalloonOverlayView are from an open source
project.
The PugBalloonItemizedOverlay class is new and is the new ItemizedOverlay
that we will use
PugAndroidActivity: Now uses PugBalloonItemizedOverlay and actually displays
game info when you tap the icon.  This will be improved to display more info
and we need to display the different icons too