Fixes NullPointerException on edit
Fixes #28. Adds a template class for the activities that display a list
of Games. Changes the way the lists were updated after an edit from a
single modification (that was very prone to throwing said
NullPointerExceptions) to completely rewriting the list. Removes the
ability to edit an owned Game from ListGameActivity due to
complications with passing all edited Games back to PugAndroidActivity,
as well as from the balloons on the map.